Generalized parenting in BGUI
-Parenting is now generalized, so any widget could parent any other
widget. Right now it's done only for windows and buttons, because right
now I don't want to add about 12 functions per widget. This is useful
for things like buttons having an image on top (add a label widget to
the button widget) or to have windows inside windows.
-Added font caching in widgets, so the font doesn't change after drawing
a widget. Everything else was remembered like color, alpha and font
alignment, but the font itself wasn't.
-Added a "zero" flag in rect and rect_offset classes and optimized
drawing using them. When a style has zero padding, then we don't need to
use draw_sprite_padded, but instead can use draw_sprite_streched. As the
padding uses 8x more vertices, then this can greatly help.
-Data elements, like groups, skins and styles are now saved seperatly.
This improves performance, as the drawing lookups don't need to go
trough these non-drawable elements. This also decreases memory
consumption, as previously one drawing element was 384 bytes, but now
it's only 144 bytes. This is because when using the unrestricted unions
the size of an element is the maximum size of the classes the union
contains. Turns out the biggest is the skin element, so data elements
weighs 392 bytes.
-Window stencil mask can now be changed. It can be done via
gui_window_set_stencil_style, where the style contains a sprite that is
going to be drawn for the stencil. Together with infinite parenting this
allows things like maps or large images inside windows with controls.
-Added d3d_stencil_continue_mask() that is used in BGUI, but it is not
enough. It seems I will have to add more stencil functions (or modify
the existing) so we can have windows cropped inside windows.
-draw_sprite_stretched_ext was actually implementing the same as
draw_sprite_stretched. So removed code duplication.