windres has flags now and compiler optimization
-For 64bit's we need to pass a flag to windres. I added this to the
e-yaml's and compiler, so it is possible to pass it.
-The compiler is now compiled with -O3 which does make the parsing
faster. This required some changes as the parser code is a mess. It has
many functions declared inline in .cpp files which is wrong. This means
that if the functions are actually inlined (as in O3) then it will trow
undefined references if other .cpp's relyed on that function being
there. Now moved a lot of that to .h, but all of it needs to be cleaned
up. There is also a lot of code duplication, like the function fc()
which loads a file into a string is implemented in about 3 or 4 .cpp
files.