New functions and fixes
-Added window_set_maximized() and window_get_maximized() on Windows. Not
sure why they weren't implemented. Also added them for xlib, but THEY
ARE NOT TESTED!!!
-64bit compile mode on Windows now uses a unique directory name
("Target-platform" in .ey). This means that you can have compiled 32bit
and 64bit ENIGMA in parallel.
-Added optimizations in BGUI - if the text of the GUI element is empty,
then a lot of code is now skipped. Extra useful with parenting like
adding icons to buttons, where buttons would have no title text and the
icons (gui_label) wouldn't have any text either.
-Added gui_continue_propagation() to BGUI. This allows the event
propagation to continue even if it would be stopped. Like if you have a
button over a window. Normally the window would not get an event,
because the button is pressed. If you want the window to get the event
you put "gui_continue_propagation" in the button callback. I use this
for things like double clicks on window titles. If you click once it
works like regular window click, but if clicked twice the button does
something different.
-Added gui_windows_group_update() which allows updating windows in
groups. It also has a second "continueProp" argument, which allows the
groups to be treated individually or together like gui_windows_update()
would.
-Added window_update_mouse() which allows calling an update for mouse_x
and mouse_y variables. Useful if you depend on them (like the BGUI
extension does), and you need to update them in the middle of a step or
draw event.
-Added window_update() which just calls ALL the events for all
instances. Note that his can cause infinite loop if called in a place
like step event. But useful in alarms, one shot events or places like
the resize() event.