Some parser fixes
-Parser uses a lot of unsafe pointers everywhere. And it doesn't check
for NULL anywhere, so there are many cases where a NULL pointer is cast
to string and a segfault occurs. Here are some basic fixes in one case,
but I'm sure this is where most of the problems lie. This fix made
crashes somewhat fewer.
-Also casting is a lot better (using  C++11 functions), so no need to
use unsafe functions like sprintf.