Changed compiler makefile so it runs from Win cmd
-Previously it called "shell find", which works on Linux and properly
configured mingw or msys installation. But if the makefile was called
from regular cmd or powershell then an error occurred, because "find" is
also a Windows program with different calling arguments.
All of this was replaced with a universal make solution (recursive
wildcard's). Now it works on Windows and should work everywhere else as
well.