Font changes, static linking and BGUI
-Added -static to both 64bit and 32bit targets. This makes all libraries
to be statically linked if possible. I use mingw that requires pthreads,
which are not statically linked by default. Instead of adding that
library individually I think this is better, as we want games to be as
self contained as possible.
-Added font_height that returns height of the maximum character.
Previously string_width("M") was required or something similar, but this
is O(1) as we already have that number.
-font_get_ functions return int instead of unsigned int. This was
especially bad because of this humorous line: "unsigned(-1)".
-string_ ... _line functions now return 0 if the line selected doesn't
actually exist. Previously it returned the length of the last valid
line.
-Now all returns in the string_width functions are "ceil()'ed", not only
the last return. I'm not sure this is correct, but Robert back in June
2014 believed they should be (a81fea0f307ec9584641c62397bf907f78628b75).
-More progress on BGUI textbox.