Additions for custom attributes
-Renamed the new d3d_model_vertex_ functions into d3d_model_add_ as it
makes more sense and doesn't conflict with other names.
-Added _add_color, _add_texcoord, _add_normal. The only different from
_float2 or _float3 is that they can be used without formats. That means
the default shader supports them.
-If model_begin(format) is done with format = -2 (which is the default)
the format will not be changed. This is useful when use several _begin
and _end functions in one model.
-Added a get_program() macro to GL3shader which will be used more later.
I found a bug that if the ID's are not defraggmented the program will
segfault as it will try to use an invalid program vector index. This
needs to be investigated further.