Surfaces, models and matrices
-Surfaces now have an option to have a stencil buffer, as well as
readable depth buffer. For a depth buffer to readable it had to be an
FBO, as Renderbuffers cannot be sampled. But then the buffer is slower,
so I ended up implementing it as a choice.
-New function surface_get_depth_texture() to get surface depth texture
if the surface has depth buffer and if its not write only.
-Added d3d_set_color_mask(r,g,b,a) function to enable/disable writing to
a color channel.
-Added d3d_stencil_enable, d3d_stencil_clear_value, d3d_stencil_mask,
d3d_stencil_clear, d3d_stencil_function and d3d_stencil_operator to have
more fine grain control of stencil buffers.
-d3d_model_ format can now be changed with  d3d_model_format. Previously
this was specified only at _begin(). This allows the same model to be
drawn with a different shader which as a different format.
-d3d_projection_stack functions now push and pop view_matrix as well.
Without it the projection stack didn't actually work.
-Small optimization in GL3modelstruct. Previously it recalculated stride
all the time, but now it is saved. It isn't as big of a deal with
default vertex format, but if you have custom one then it requires a for
loop with a switch.
-Fixed a potential SEGFAULT with surface_getpixel_ext.
-gui_slider_set_value now properly checks bounds.