Fix GL1 and fixes/optimizations for GL3
-GL1 now has all the same changes in matrix functions so it would
compile. The same bug fixes apply too.
-d3d_projection_stack_clear didn't clear the view_matrix stack.
-d3d_model_get_stride returns pre-calculated stride instead of
recalculation. This should be a lot faster in many cases, as this is
called every time a texture batch check happens. Basically every time a
draw function is used.
-Fixed surface error in debug mode.
-get_texture in GL3textures is now a macro and shows an error in debug
mode.
-Fixed a tile bug in GL3 that meant the wrong texture is used. Screwed
up rendering with tiles. Thanks to rcobra's example for finding this
bug.
-Changed the order of Authors in OpenGL3 about section, so people know
who to contact first in case of a bug.