Optimizations and new transformation functions
-Matrix4::init_camera_transform is now slightly faster, because matrix
mult was replace with 3 dot products.
-Added Matrix4::init_look_at_transform which is used by
d3d_transform_set_look_at and d3d_transform_add_look_at in GL1 and GL3.
This rotates an object to face a point. Very useful.
-Added d3d_transformation_get_mv and d3d_transformation_get_mvp that
returns model_view_matrix and model_view_projection_matrix respectivly.
Useful in cases like shadowmapping where we use projection function to
set a camera and then we need MVP in a shader later.
-Math function lerp() now uses C++11 std::fma function. It is basically
as fast (or faster) than what was done previously, but it returns
correct results when t == 1 or when x == y.