BGUI additions and get_texture macro fix
-get_texture didn't handle texid==-1 well. Fixed it.
-gui_windows_update and gui_windows_group_update now return if the event
propagation was stopped. That basically means the mouse was over a
window or any widget bound to the window. Useful if you want to combine
GUI with mouse events that are not tied to GUI, like RTS units shouldn't
go to a waypoint if the mouse was actually clicked on a window.
Previously there wasn't an easy way to check this.
-Many textbox improvements:
--ctrl+a now select the whole textbox content.
--Added settable line limit and chars per line limit. Useful if you want
to make a "text field" with 1 line which can have up to 200 chars
instead of an unlimited "text area" for example.
--Numeric only support is a lot better. It now supports anything a
std::strtod can, like scientific e notation. There is a problem though,
that the number must be numeric at all times, so you cannot actually
type "10e5" from left to right as "10e" is not a number and so it won't
go past "10". Open to suggestions on how to fix it. Maybe buffer the
string until deselect is called.
--Fixed position select with mouse that was wrong before. This is
because string_width() returns a rounded number, but I need more
precision. Added string_char_width() to compensate, but I suggest making
string_width/string_height functions to return a double. In which case
string_char_width() wouldn't be needed here.