Added internal versions of glsl_ functions
-Added internal versions of some glsl_uniform and glsl_attribute
functions that skip batch flush. A few commits ago I added flushing to
these functions as they change the drawing state, but this created
recursion in modelStruct::Draw, as batch flush called it, and it in turn
calls glsl_ functions. This should be also faster and later I will
remove some checks as well.
-glsl_get_uniform_location and glsl_get_attribute_location errors are
now printed only in Debug mode.
-Replaced GetStride() with precalculated stride. Should be faster.