Fixed small bug with directory_rename() / _copy()
prior to this commit, directory_rename() and directory_copy() wouldn't work if you tried to copy or rename a directory to the same parent directory where the source directory is a sub-string of the destination directory. For example, rename/copy "folder" to "folder2" wouldn't work. This commit fixes that so it's no longer an issue.