Fixes for Windows and cleanup (#1076)
1) Cleaned up all of the about files to have the correct specificity.
Windows should now be selected as the correct default platform when on
Windows, and OpenGL1.1 should be the default graphics everywhere, since
again, we consider it most stable.
2) Reverted @fundies change to the ENIGMASystem/SHELL/Makefile because
what @JoshDreamland had before was correct. This didn't even compile and
broke Windows which was reported in #1033 that we couldn't figure out.
3) Added follow_object to roomsystem.h so the graphics systems would
compile again.
4) Also added missing end of lines where appropriate.