Update lcov directory; add failure if lcov fails.
We recently changed ENIGMA's output directory, but did not update the lcov invocation to point to the new default location. This commit makes that update, hedges against changes to the configuration, and reports a GTest failure if coverage fails. This should fix coverage and prevent it from being accidentally broken in the future.