Fix timeline actions
It was always action_timeline_set but I have evidence I will show in the PR that I think YYG dropped the red ball in GM8. We can work around it by using a simple define, but I chose to rename action_set_timeline to be action_timeline_set because that was the original name in GM7 and earlier.