WINDOWSmain.cpp now supports UTF-8
Everything except the CreateWindow() which is unnecessary in this case and windowsystem_write_exename() which still needs to get done but idk how to make it support UTF-8. If windowsystem_write_exename() is made to support UTF-8, there will no longer be an error message when you try to run your game inside a path with special characters.