fix for get_open/save_filename_ext()
I noticed i replaced the quotes with quotes using escaping, before checking the existence of the path with stat(), and the existence of the file with access(), this means if you used quotes in your string for the fname or dir aguments, the file and path will not be recognized. To fix this, I called string_replace_all() on fname_or_dir after stat() and access() instead.

I also reorganized the functions around and reformatted the way the FileFilter class was written, without changing it's behavior.