Extend Vertex Buffers With User Index Buffer Features (#1283)
Additionally functionality for clearing/unfreezing a vertex buffer and specifying their contents more expressively to enable tiles, models, and other functions to be rewritten generically and more clearly. Index buffers also added to extend the capabilities of the vertex buffers for more advanced geometry rendering, such as level of detail or geometry mipmapping for 3D terrains.