Update View Following Before Draw (#1430)
Loop the views once to update the object following before looping the views to fire draw events. This ensures that all of the views are updated in the first frame and the user can access views other than the current one. Also expands the room transition test to verify this behavior in the first room's draw event before switching rooms and to prevent future regression.