Compiler: use GameData instead of EnigmaStruct.
Refactor the compiler to hand around GameData, a tuple of proto messages, instead of EnigmaStruct. This allows IDEs to pass protos in instead of building a shared-memory system around EnigmaStruct.

Also rewrite pieces of the compiler that have not aged well. Delete some redundant and unused code. Remove some duplication and global state.