Compiler Cleanup: Remove more global state, fix alarms
Some global state wasn't being updated because I cut a bunch of it out, deleted a redundant copy of some code, and lost a little information in the process. The alarms fix was really just updating the `namespace_enigma` global. Instead of doing that, I've removed the `namespace_enigma` global. Good fucking riddance.

There's still some more global state lying around, and ENIGMA is passing around JDI definition pointers willy-nilly, which it should NOT be doing. That's entirely my fault, but it still hurts the same.

In the future, we'll want to pull out the JDI definitions from the compiler. But the only occurrences of JDI-isms, now, are within `lang_CPP`, the syntax checker, and the code translation / "linking" module. The syntax checker and code translation unit are going to be replaced by a tree builder and pretty printer, respectively. When this occurs, I intend to abstract away the process of symbol definition (checking if a function exists, checking if it's variadic, etc). The linking module will only need minor tweaks. It's actually pretty brilliant for young me; I'm impressed.