Make Mac dialogs use *.icns file set by Info.plist
Before this pull request, the dialogs that display an icon would show the default icon for app bundles that have no icon set. This change makes sure the dialog recognizes and uses the icon set for the app bundle by the Info.plist if the file happens to exist, otherwise the fall back on the previous behavior.