KDialog Now Supports Multiple File Filters!!!!!!!!
I had no idea this was possible but after many hours of research I discovered each file filter pair is supposed to be separated by a new line character in order for multiple file filters to be supported in KDialog. There are also a few other fixes I added to KDialog including the ability to set the default directory and file name for the file dialogs.