TinyFileDialogs improvements (#1525)
* Make Mac dialogs use *.icns file set by Info.plist

Before this pull request, the dialogs that display an icon would show the default icon for app bundles that have no icon set. This change makes sure the dialog recognizes and uses the icon set for the app bundle by the Info.plist if the file happens to exist, otherwise the fall back on the previous behavior.

* KDialog Now Supports Multiple File Filters!!!!!!!!

I had no idea this was possible but after many hours of research I discovered each file filter pair is supposed to be separated by a new line character in order for multiple file filters to be supported in KDialog. There are also a few other fixes I added to KDialog including the ability to set the default directory and file name for the file dialogs.