Made TFD easier to read (#1529)
Previously the inconsistent indention used in in TFD was one big mess and made it a tiny bit difficult to read and looked very sloppy. This should fix it for the most part, if not completely.

I kinda wanted to go even further and make this file match the rest of ENIGMA's intended intention and syntax practices, but I was too afraid I'd accidentally mess up the file somewhere. Maybe some day...