Fix Remaining Healthbar Bugs (#1539)
This pull request fixes some remaining issues with the healthbar action and function. The first issue is that the healthbar function was not drawing a border around the background color when the background is shown. The second issue is that the healthbar action function was not mapping the back color parameter to the color constants like the front color parameter. The none option for back color is handled specially by using its value as an argument to the show border parameter. Special thanks to @ProtoLink for helping me fix this, I ended up having to recreate the changes though since he lacks reliable internet to submit the changes himself.