Screen Copy Stopgaps (#1540)
This is a recreation of #1522 that has stopgaps for the other sprite and background function and for OpenGL3. This pull request is basically a temporary stopgap for #1423 until I have more time to properly cleanup all of the screen copying. Special thanks to @ProtoLink for helping me fix this issue!