Delouse the CI (#1541)
The CI script inverted the check for the make parallelism being unspecified, and so was not copying it into `MAKE_JOBS`. This check was pointless even if it worked, because if `$2` was unset, `$MAKE_JOBS` would remain unset. Consequently, `make -j$MAKE_JOBS` evaluated to `make -j`, which means "Please saturate my machine with GCC processes." It was not `gcc` spinning up a million `cc1plus` instances; it was `make` spinning up a million `gcc`s, which then `exec`'d `cc1plus`.

Travis' instability is still a little wondrous, but I guess we can continue ignoring that until it becomes a problem again in the future.