cross-platform abort/retry/ignore show_error() (#1569)
Adds the M$ standard abort/retry/ignore error message buttons to show_error() but doing so on all platforms and widget systems. when fatal is true all the buttons abort the game where as only the abort button does that during non-fatal. I'm planning to get rid of the retry/ignore buttons for fatal and retry for non-fatal as they serve no purpose in these contexts, I'm just waiting for KDialog authors to add the --ok-label command line option so that this stuff will be possible for all of our widget systems. It is in their plans to do it, confirmed on their GitHub repository notes. This will allow me to change the single "OK" button to "Abort" for KDialog's fatal error message box, then I'll update all the widgets.

Also reordered the Objective-C function arguments to match the C++ ones in Cocoa Widgets.