Add Direct3D11 Support for MSAA (#1579)
Very simple, we just release the render target, its associated depth buffer, and the swap chain then set the desired samples and recreate the swap chain and render target. The query for the sample levels supported is basically the same as D3D9 but with an additional requirement to check the quality levels. Finally, I also changed the format to have 2-space indentation like the rest of the engine to make it easier to read for myself.