Patch for breaking protoc change.
The Protocol Buffer compiler, protoc, has started appending `final` to generated class definitions as of 3.7. Because of this, our extension classes will no longer build. This removes our dependency on the ability to inherit proto classes, at the cost of using `->` instead of `.` to refer to proto fields from the Data classes.