Render State Centralization (#1640)
This is a cleanup and centralization of all the redundant state definitions in each of the graphics systems. This covers d3d state, blend state, and color state. Instead of having all the sources, even in the same graphics system, externally link the current color, they can include the proper general header. The getter functions were also removed because GM never had them and we actually don't have a good explanation for them.