Generic Texture Objects (#1643)
This pull request is a friend of #1640 and involves me splitting up the redundant and duplicate texture structs. The generic part of them I moved to `General/` to create a generic `enigma::Texture` object. This class is an abstract base class which can not and should not be constructed directly. Each backend then uses virtual inheritance to store the native peer texture object with the generic texture object. This requires the backends continue creating the textures, for now, so they have the correct type. This allowed moving several of the texture functions to general since they only operate on the state provided in the general texture object. This includes getter functions for the texture dimensions which are still provided by GM to this day.

I also created the directory `Graphics_Systems/OpenGL` for common code between the GL systems. I moved every generic GL code there and out of `General/` to keep the directories organized. The problem with having GL code in `General/` before is that it was accidentally creeping into D3D and None.