Direct3D9 Fatal Display Reset (#1645)
This one is pretty obvious, I am switching from `MessageBox` to `show_error` so that I can force the game closed whenever a display reset is not successful in the Direct3D9 bridge. It's important to do this because this is not a scenario in which we can continue running the game. It does not matter if the display reset is caused by us or the result of a lost device. I tested to ensure this does what I think. I put myself in the managed Direct3D9, commented out releasing of the surfaces on lost device, and noticed the `MessageBox` version segfaults after showing the reset failed message. The new version shows our standard error dialog and then forces the game closed with `Game returned 0`. I also removed some unused headers and namespaces while I was in here.