Direct3D9 State Reset (#1647)
When we reset the Direct3D9 device, we only need to restore our state cache if we are using the normal, managed Direct3D 9.0 device. A 9Ex device does not lose state information. So this pull request simply adds a condition to our Reset helper to only restore the state when using the managed device.