Remove DlgMod visibility due to being broken
with dll calling broken and no progress being made with that we can remove this as a temporary measure to not confuse users. we already have a broken GTK+ widget set, we don't need a any other broken widgets sets. there is no point, its broken until we fix dll calling, and people prefer not having a dll and will go with windows widgets anyway