Draw Batch & Vertex Submit Mixing Fix (#1704)
A state flush implies you are going to draw something anyway, so it doesn't matter if the state is actually dirty with respect to flushing the batch, so we always flush the batch when a state flush is requested. This fixes user submission of vertex buffers that occurs directly after a batch is completed (e.g, draw_rectangle) when no state change has occurred between.