Fix using string as type/function.
When we moved exclusively to namespace `enigma_user`, we didn't move toString there, so the function is no longer allowed in EDL. Problem is, `string()` was a macro to it, so that was necessarily broken.

The best option for now is to remove the macro and use just the type in EDL. EDL will not force the std::string constructor to succeed on objects passed to the method, so GCC can silently fix it in the preprocessor for the time being.