center linux game on screen at startup
Robert - If you close this Josh might have to revert your privilege level back to a normal contributor, becuase josh told me to make this pr. Instead of closing it to be obnoxious, request changes like a good boy, and maybe santa clause won't give you coal this year.

this effects all platforms and on windows and mac it is redundant, but oh well, it is required to initialize_everything() first, to determine game window size, before linux game can be centered properly, rather than opening in the corner of the screen or w/e.