AssetArray ADT for Resource Storage (#1767)
Provide dense, cache-efficient resource storage abstraction. Switch DirectSound to use it for holding sound resources. This will enable us to switch other systems soon and eliminate duplicate logic and semantics.

Behavior provided by AssetArray is modeled on GM8.1 and should be expected to produce similar results. We may be able to swap the use of vector with a custom slab allocator at a later point in time.