AssetArray Move Fixes (#1800)
Fix move semantics of the AssetArray class by taking non-const rvalue references. It's not actually possible to move from const rvalue references, and it just falls back to the copy constructor. This fixes audio in the DirectSound system.

Use a COM smart pointer for sound buffer in DirectSound. This allows the sound to be safely and efficiently moved while also allowing it to be constructed and deleted safely, even without AssetArray managing it.