Asset Exists Iterator (#1814)
Provide a special AssetArray iterator that allows you to loop over only existing assets in a range-based for loop. This will simplify existing functions and correct some logic mistakes in a few places. Reuse of ids a la GMSv1.4, a skip list, or a segment tree can be investigated at a later point in time.