Finish the proto-yaml reader.
It is now capable of reading the new events.ey. It offers generic support for deserializing YAML to proto, including some special behaviors around repeated fields.

Specifically, if a proto message is a key-value pair, or more generally, has a key (name, id, key) field, a map can be used in YAML instead of a sequence specifying IDs. A sequence can also be used in these cases. If there is only one other field than the ID, proto-yaml allows it to be specified directly as the value, completely removing the key/value field names from the YAML.