Fix tests some more.
Remove C++17 dependencies.
Clean up some crap I never ended up needing.
Make the new event consolidation logic in write_defragged_events consistent with one of the MANY places it is duplicated needlessly in write_object_data.
Long term, we need to rename the UsedEventIndex to DefraggedEvents so that people know what the fuck that is, and just hand that struct over to the write_object_data logic.

Fuck.