Fix SDL Without Breaking Eveything Else
it was a one-liner change fundies told me to do a while back. He knew all along this would fix it because he gave me the idea. Yet he went ahead and made a pr with tons of my changes removed that actually benefited the project hugely just to piss me off.