Default Makefile Consistency (#1906)
Add the underscore we forgot in #1905 to OBJ_DIR in the default makefile so all of the top-level makefiles are consistent with respect to OBJ_DIR usage.

Note: SHELL makefiles still use the old OBJDIR without underscore because fundies did not touch those. However, they could be cleaned up at a later time by also using shared makefiles in a similar fashion.