Add More Async Dialogs
I've decided to add support for async dialogs for every dialog function, as well as fix a segfault with the macOS async dialogs. On macOS, a window cannot be created in any thread other than the main one; so as a workaround, you may download separately my command line program as a dependency, and put it ino the "Resources" folder of your application bundle, to get it all working:

https://github.com/time-killer-games/dlgmod

Windows or any other UI elements attempted to be created or modified outside the main thread will cause a segfault, so this is 100% necessary. There are other ways to achieve async dialogs that don't require launching a child process or multi-threading, but this method is much simpler to incorporate with our current async dialog extension implementation due to our use of threading.