fix borderless toggle (again) (#1934)
ok, so my last fix wasn't so great. I wanted to keep the client area of the window to not change position on screen when toggling the border, and since every window manager handles this differently, i decided to just revert this back to how it originally was, and forget about that all together.