these functions were written by a dummy
the fname argument needs quotes around it to accept spaces and it also needs a space between that and the args argument to the fname and first argument aren't fused together like Frankenstein's monster. Whoever wrote these functions initially was on crack. At least now, it works how it's supposed to, like it does on Windows, thanks to me. I can't believe this has been broken for the 10+ years ENIGMA has been around lol