Game Compilation Interrupt (#2068)
Add a new DLL procedure that flags ENIGMA to send a CTRL+C SIGINT to any child process groups, such as make or the game, it has spawned during the compilation phase. This facilitates an IDE or RPC to provide a stop button to the user which can be enabled as soon as the ide_dia_clear callback is fired.