Finish EgmEvents (#2104)
* Check in a new events.ey file, and a proto that defines it.

Start work on a generalized, somewhat lax YAML to Proto reader.

* Finish the proto-yaml reader.

It is now capable of reading the new events.ey. It offers generic support for deserializing YAML to proto, including some special behaviors around repeated fields.

Specifically, if a proto message is a key-value pair, or more generally, has a key (name, id, key) field, a map can be used in YAML instead of a sequence specifying IDs. A sequence can also be used in these cases. If there is only one other field than the ID, proto-yaml allows it to be specified directly as the value, completely removing the key/value field names from the YAML.

* Break everything

* Plug the new event reader into the compiler. This kills the crab.

To-do:

- Test harness needs to be hooked in instead of being run ad-hoc with a gross makefile.
- There's an infinite loop somewhere; testing isn't finding it.
- Super check assumptions are wrong; super checks *can* use parameters. Compiler needs to be refactored further around how it treats super checks.
- Case insensitivity in param lookup maps. Currently retrieval of "key" will fail because it's called "Key" in the parameter value map.
- Need a method like ParamSubst that swaps in parameter names instead of argument spellings.
- Tons of migration work to get the whole compiler on the new system. Until then, resource names can't be used in filenames.

* josh helpless w/o me

* Finish adapting the new events file.

As far as I know, this fixes the new event system. Good as old.

Still a lot of work to be done migrating to this new system, but this is a good first step.

Minor to-dos remain:

- Mappings are still case-sensitive. In fact, the engine is case-sensitive, so we have to ToLower the identifying bits of function names.
- Human names for EventDescriptor structs don't replace parameter placeholders with the parameter name, so that's a bit tacky.

Apart from that, I think we're in business.

* fix typo; s/events.res/events.ey

* Fix makefile merges

* Move -lgtest after -lgtest-main

* Remove CommandLine/emake-tests/Makefile and old target.

No longer needed, and conflicting with main rules.

* Fix tests

* Fix tests some more.

Remove C++17 dependencies.
Clean up some crap I never ended up needing.
Make the new event consolidation logic in write_defragged_events consistent with one of the MANY places it is duplicated needlessly in write_object_data.
Long term, we need to rename the UsedEventIndex to DefraggedEvents so that people know what the fuck that is, and just hand that struct over to the write_object_data logic.

Fuck.

* Let Android build depend on libyaml-cpp-dev.

* Build yaml-cpp anew for Android.

* Make ymal-cpp build on Trusty, maybe

* Fix NoMoreHealth, delete events_localless.res

* Pull in EGM changes from rebase branch.

* Freshen. Merge master. Fix conflicts. Fix build.

* Pull in fundies' build fixes for gm2egm.
Thanks, fundies.

* Use proper EGM event filenames.

Prior to this change, EGM uses integer values in all parameters. After this change, only "integer"-type events will use integers in filenames, or else errors will be reported.

* Fix event ordering at the cost of explicitly naming the ID field of events.

* Whatever

* Forbid copy construction of d/varray

* Only generate subcheck functions when needed. Use function body.

* Smear polishing compound on modified areas

* idk, rewrite some more shit

This is exactly like that jockstrap incident, only now we don't have sorlok to dig the holes

Also I'm sleepy and I need to rename like all of this shit later

Did I mention it still can't correctly build a game? I keep uprooting ugly shit. Latest victim: default event storage. They need to live in the  event_parent, but have iterators registered in each root object. Think that's accomplished.

Biggest remaining problem seems to be subchecks, which are marked virtual in the event parent for some fucked-up reason. Real problem is that they're not being implemented in objects that declare them . Shouldn't be hard to fix. They have to stay functions, though, because they need access to locals.

* Smear some more shit around; fix everything?

Seems to work. It's possible `draw_gui` is broken. I don't care to check. It shouldn't be.

As for what else might be broken... life is like a box of chocolates. Sometimes that box is from Russel Stover. Ew.

* Fix includes in ALsystem.h

* Always install newer yaml-cpp for CI.

Old one is fucking ancient and doesn't have Node::Mark().

* Update Main.cpp

* can't subdir protos cause cmake gay

* recuse glob

* add proto yaml to cmake

* rgm changes needed

* Update azure-pipelines.yml

* add cmake dep

* Fix emake tests

* Update connected-components.cpp

* Fix case typos in events.ey

* remove dup File2String

* fix windows shared makefile link of protocols

* Make the hack slightly less hacky and actually work

* fix cmake links

* Removes duplicate legacyEvent2egm event. Moves event conversion from compile to project load for libegm. Now clears legacy_events on load. This fixes egm writting but egm loading still seems to be broken.

* actually read events from egm

* write gmk sprites & backgrounds with logical names so egm conversion is neater

* typo

* fix room creation code

* windows bad

* write better names for all resources

* need to pass size for backgrounds

* proper locals event fix

* windows again

* Update gmk.cpp

* Update events.ey

* Update azure-pipelines.yml

* Convert DnD IDs

* Add event dispatchers, remove prefix/suffix code.

* Fix health.

The codegen is a little undisciplined in that subcheck methods are always emitted for objects, even if they're never used. The cases in which they're used are a bit complicated, so I'm comfortable leaving that alone for right now.

* Delete suffix from Out of Health.

It doesn't need one. I don't know how shit like this ends up in the codebase. If someone's game absolutely depends on running out of health every frame, add a dispatcher.

Co-authored-by: Greg Williamson <cheeseboy16@gmail.com>
Co-authored-by: Greg Williamson <fundies@no-reply.github.com>
Co-authored-by: Greg Williamson <fundies@users.noreply.github.com>
Co-authored-by: Robert Colton <robertbcolton@gmail.com>
Co-authored-by: Greg Williamson <gwill6@uis.edu>